move everything out of cli module
> Insert changelog entry or delete this section.
