+++ python_modules/dagster/dagster/components/list/__init__.py
@@ -1,0 +1,5 @@
+from dagster.components.list.list import (
+    list_all_components_schema as list_all_components_schema,
+    list_definitions as list_definitions,
+    list_plugins as list_plugins,
+)
