+++ python_modules/dagster/dagster/components/list/list.py
@@ -78,7 +78,7 @@ def list_all_components_schema(
     return TypeAdapter(union_type).json_schema()
 
 
-def list_definitions_impl(
+def list_definitions(
     location: Optional[str],
     **other_opts: object,
 ) -> list[DgDefinitionMetadata]:
